requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=1.1.0
groq>=0.30.0
python-dotenv>=1.0.0
